refactor(tui): extract conversation rendering component
